remove hardcoded paths
deletetemps will now offer two options: delete only windows update directories that follow a specific format, OR delete all directories within the TEMP directory.

The TEMP directory used to be hardcoded. This was not OK as it required users to modify code. We are now fetching the TEMP directory from the windows environmental variable "TEMP", as we should have done all along.